feat : 메인에서 예산 플랜 없을 때의 경우 추가
